Make sure FWEventItemsManager resets himself on clear
Before this FWEventItemsManager::clearItems was not resetting the internal
vector m_items, but simply zeroing the old elements and appending the new ones.
I changed it so that while the zeroing still happens when removing any given
item, when removing all of them (for example on configuration reload) the
holding vector is properly cleaned.

Notice that FWModelChangeManager now listens to the the goingToClearItems_
signal and resets itself accordingly.